move SearchBar to separate component
